Added Layout/BeginEndAlignment
Enforces this cop to mimic Layout/EndAlignment behavior for `if`,
`while`, `case` statements. Example:

    result = if true
             end

    result = begin
             end

While the results of these statements should never be assigned to a
variable, at least this will make it hurt more visually so one can
extract to a method and assign the results of the method to a variable
for improved readability.